Add CodeQL Advanced (#9)
This pull request introduces a new CodeQL workflow configuration to
enhance code security analysis for the repository. The most significant
change is the addition of a `.github/workflows/codeql.yml` file that
defines a matrix-based CodeQL analysis job for multiple languages,
including Swift and GitHub Actions.

### New CodeQL Workflow Configuration:

*
[`.github/workflows/codeql.yml`](diffhunk://#diff-12783128521e452af0cfac94b99b8d250413c516ec71fe6d97dbea666ff7ba27R1-R44):
Added a new workflow named "CodeQL Advanced" that triggers on pushes and
pull requests to the `main` branch. It uses a matrix strategy to analyze
multiple languages (`actions` and `swift`) with tailored steps for each
language. For Swift, it includes custom steps for creating a secrets
file, selecting an Xcode version, and building the project before
performing the analysis.